requirements.txt
aiohttp==3.9.0
aiosignal==1.3.1
anyio==3.7.1
APScheduler==3.6.3
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
beautifulsoup4==4.12.2
black==23.11.0
cachetools==4.2.2
certifi==2023.11.17
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.5.14
distro==1.8.0
docx==0.2.4
docx2txt==0.8
emoji==2.8.0
exceptiongroup==1.2.0
faiss-cpu==1.7.4
filetype==1.2.0
frozenlist==1.4.0
google==3.0.0
google-ai-generativelanguage==0.3.3
google-api-core==2.14.0
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-generativeai==0.2.2
googleapis-common-protos==1.61.0
greenlet==3.0.1
grpcio==1.59.3
grpcio-status==1.59.3
h11==0.14.0
httpcore==1.0.2
httplib2==0.22.0
httpx==0.25.1
idna==3.4
joblib==1.3.2
langchain==0.0.190
langdetect==1.0.9
lxml==4.9.3
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
nltk==3.8.1
numexpr==2.8.7
numpy==1.26.2
openai==0.28.0
openapi-schema-pydantic==1.2.4
packaging==23.2
pathspec==0.11.2
Pillow==10.1.0
platformdirs==4.0.0
proto-plus==1.22.3
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==1.10.13
pyparsing==3.1.1
pypdf==3.17.1
python-dotenv==1.0.0
python-iso639==2023.6.15
python-magic==0.4.27
python-telegram-bot==13.7
pytz==2023.3.post1
PyYAML==6.0.1
rapidfuzz==3.5.2
regex==2023.10.3
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.23
tabulate==0.9.0
tenacity==8.2.3
tiktoken==0.5.1
tomli==2.0.1
tornado==6.3.3
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.8.0
tzlocal==5.2
unstructured==0.11.0
uritemplate==4.1.1
urllib3==2.1.0
wrapt==1.16.0
yarl==1.9.3